moved parsing to worker thread
